+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+🚨 This release contains some breaking changes. See the Removed section for more information. 🚨
+
 ### Added
 
 - Added `get_component_names_used_in_template` method to ComponentRegistry to access component names used in a template
@@ -39,7 +41,9 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 ### Removed
 
 - Removed automatic component scanning at application startup for faster initialization
+- Removed `ENABLE_AUTO_CONFIG` setting and autoconfiguration functionality (moved to django-bird-autoconf plugin)
 - **Internal**: Removed `ComponentRegistry.discover_components` method.
+- **Internal**: Removed `AutoConfigurator` class used for automatic setting configuration
 
 ## [0.16.2]